pyproject.toml
[project]
name = "bara-chat"
version = "0.1.0"
description = "Local privacy-focused chat application"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "sqlmodel>=0.0.14",
    "PySide6>=6.6.0",
    "aiortc>=1.6.0",
    "websockets>=12.0",
    "pynacl>=1.6.1",
    "Pillow>=10.2.0",
    "pyjwt>=2.8.0",
    "pycryptodome>=3.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

